refactor(notifications): NotificationRuleEditor 를 도메인 모듈로 분할
1,156줄 단일 파일을 5개 파일로 분할했다 — NotificationRuleEditor 734 ·
ChannelConfigForm 284 · api 72 · types 61 · constants 40.
JSX·props·훅 호출 순서를 한 글자도 바꾸지 않았다(rules-of-hooks 보존).

계획서 인벤토리가 "타입·상수 164줄" 로 묶은 구간이 실제로는 세 층이었다 —
타입(35-91) · 상수(97-128) · **API 호출 레이어(133-193, 함수 8종)**.
API 함수를 메인이 10회·ChannelConfigForm 이 2회 쓰므로 api.ts 로 분리하면
양쪽이 공유하는 자연스러운 경계가 된다. 인벤토리보다 한 축 더 나눴다.

ChannelConfigForm 은 index.ts 가 재노출하지 않는다 — 소비자가 메인 컴포넌트
하나뿐이고(실측), claude-sessions/TranscriptViewer 선례를 따른다. 원본에서
지역 함수였으므로 export 를 추가했다.

패키지 승격 단계는 불필요했다. components/notifications/ 디렉토리와 index.ts
가 이미 있어 파일만 나누면 됐다.

tsc 의 noUnusedLocals 가 grep 심볼 카운트의 오차를 3건 잡았다(Webhook·
Settings·testChannel 이 "쓰인다"고 세어졌으나 실제 미사용). B2 에서 ruff 가
한 역할을 프론트에서는 tsc 가 한다 — import 이전의 진실원은 grep 이 아니다.

게이트: tsc 0 · ESLint 0 · vitest 61 passed · build 성공.
